requirements.txt
# ============================================================================
# REQUIREMENTS - Temporal Action Detection System
# ============================================================================
# Sistema profesional de detección temporal de lenguaje de señas
# Basado en arquitecturas Transformer estado del arte (2023-2025)
#
# Instalación:
#   1. pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#   2. pip install -r requirements.txt
# ============================================================================

# ===== CORE DEEP LEARNING =====
# PyTorch (instalar manualmente con CUDA)
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# Transformers y modelos pre-entrenados
transformers>=4.30.0
timm>=0.9.0  # PyTorch Image Models (incluye Video Swin)
einops>=0.6.1  # Operaciones tensoriales elegantes

# ===== COMPUTER VISION =====
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1  # Augmentations avanzadas
imgaug>=0.4.0

# ===== DATA SCIENCE & PROCESSING =====
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ===== VISUALIZATION =====
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0  # Visualizaciones interactivas
tensorboard>=2.13.0

# ===== VIDEO PROCESSING =====
av>=10.0.0  # PyAV para lectura eficiente de video
decord>=0.6.0  # Lectura rápida de video
moviepy>=1.0.3  # Edición de video

# ===== METRICS & EVALUATION =====
scikit-video>=1.1.11  # Métricas de video
torchmetrics>=1.0.0  # Métricas modernas PyTorch

# ===== CONFIGURATION & LOGGING =====
pyyaml>=6.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.5.0  # Terminal output bonito
loguru>=0.7.0  # Logging avanzado

# ===== PROGRESS & MONITORING =====
tqdm>=4.65.0
wandb>=0.15.0  # Weights & Biases (opcional)

# ===== OPTIMIZATION & ACCELERATION =====
ninja>=1.11.0  # Compilación C++ rápida
flash-attn>=2.0.0  # Flash Attention (opcional, requiere compilación)

# ===== UTILITIES =====
python-dateutil>=2.8.2
psutil>=5.9.0  # Monitoreo de recursos
py-cpuinfo>=9.0.0  # Info de CPU

# ===== DEVELOPMENT (opcional) =====
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.1.0
# mypy>=1.5.0

# ===== GPU MONITORING (opcional) =====
gpustat>=1.1
nvidia-ml-py3>=7.352.0
pynvml>=11.5.0

# ===== DATASET MANAGEMENT =====
h5py>=3.9.0  # Almacenamiento eficiente
zarr>=2.16.0  # Arrays comprimidos
lmdb>=1.4.1  # Database rápida

# ===== ANNOTATION TOOLS (opcional) =====
# labelme>=5.2.0  # Anotación de imágenes/videos
# vatic>=0.1.0  # Anotación de video

# ===== ACCELERATION LIBRARIES =====
numba>=0.57.0  # JIT compilation para Python
# cupy-cuda12x>=12.0.0  # NumPy en GPU (opcional)

# ============================================================================
# NOTAS IMPORTANTES
# ============================================================================
#
# 1. PyTorch con CUDA:
#    Instalar manualmente ANTES de este requirements:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
# 2. Flash Attention (opcional, mejora velocidad):
#    Requiere compilación desde source:
#    pip install flash-attn --no-build-isolation
#    Solo funciona en GPUs con Compute Capability ≥7.5 (RTX 20xx+)
#
# 3. Verificar instalación:
#    python -c "import torch; print(f'PyTorch: {torch.__version__}')"
#    python -c "import torch; print(f'CUDA: {torch.cuda.is_available()}')"
#    python -c "import transformers; print(f'Transformers: {transformers.__version__}')"
#
# 4. Requisitos de Hardware:
#    - GPU: NVIDIA con ≥12GB VRAM (recomendado: ≥24GB)
#    - RAM: ≥32GB (recomendado: ≥64GB)
#    - Storage: ≥100GB SSD
#
# 5. Modelos Pre-entrenados:
#    Los modelos se descargarán automáticamente de Hugging Face:
#    - Video Swin Transformer: ~400MB
#    - TimeSformer: ~350MB
#    Total: ~5-10GB con cache
#
# 6. Alternativas ligeras:
#    Si tienes GPU más pequeña (<12GB), considera:
#    - Reducir batch_size en config.yaml
#    - Usar backbone 'slowfast' en vez de 'video_swin'
#    - Reducir clip_length a 32 frames
#
# 7. CPU-only (NO RECOMENDADO para este sistema):
#    Si solo tienes CPU, instala:
#    pip install torch torchvision torchaudio
#    Pero el entrenamiento será extremadamente lento (semanas vs días)
#
# ============================================================================
